make UpdateRequest a subclass of UpdateRequestHandler
